Always query directly for build for `circleci-show`
Previously relied on the JSON returned for the list of recent builds to
show the most recent build when called without arguments; however, this
JSON does not include the steps. Its unclear if it used to do this.

Instead make it more consistent with the other subcommands that query
the most recent builds to get the build number when called with no args
followed by their actual API call.